Implement support for MSVC Z7 flag
When the "-Zi" flag is given (external pdb) on the command line
without an explicit path given by "-Fd" and it is then followed
by a "-Z7" (internal pdb), the last flag is ignored by sccache
and so it gives error because an external pdb in a shared folder
is not supported.
This is incorrect because "-Z7" overrides the previous option
and the debug symbols are actually internal, which is supported.

Also update to the latest zip-rs package version.